🗺️ Atlas: Extract TxVisibilityManager to core
